requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and async
httpx==0.26.0
aiohttp==3.9.1

# LangChain ecosystem - stable compatible versions
langchain==0.1.17
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-core==0.1.52
openai==1.25.0
instructor==1.11.3

# ML dependencies - CPU optimized versions
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.8.0
transformers==4.36.2
sentence-transformers==2.2.2
tiktoken>=0.7.0
together
accelerate
chromadb==0.4.18

# Data processing - lighter versions
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2

# Utilities
jinja2==3.1.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7
redis==5.0.1
structlog==23.2.0

# Web search for current information
tavily-python==0.3.0

# Monitoring - minimal versions
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Remove heavy plotting libraries for deployment
# matplotlib>=3.5.0
# seaborn>=0.12.0
# plotly>=5.0.0
# statsmodels>=0.14.0
